requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & Storage
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.1
minio==7.2.0
boto3==1.34.0
botocore==1.34.0

# Message Queue
pika==1.3.2
celery==5.3.4

# Analytics & Vector DB
clickhouse-driver==0.2.6
qdrant-client==1.7.0

# ML & AI
openai==1.3.7
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.4

# # Orchestration
# apache-airflow==2.8.0
# apache-airflow-providers-postgres==5.8.0

# Email
sendgrid==6.10.0
jinja2==3.1.2

# Monitoring & Utils
prometheus-client==0.19.0
structlog==23.2.0
python-multipart==0.0.6
httpx==0.25.2
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1
